requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
piexif>=1.1.3
